+++ src/containers/TransferCurrency/ConfirmationPage.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react'
+
+export const ConfirmationPagePresentation: React.SFC = () => (
+  <div>Confirmation Page</div>
+)
+
+export { ConfirmationPagePresentation as ConfirmationPage }
